OCR_Paper/OCR_Paper-main/requirements.txt
gunicorn==21.2.0
Flask==2.3.2
python-dotenv==1.0.0
apscheduler==3.10.3
jwcrypto==1.5.0
dataclasses-json==0.5.14
opencv-python==4.6.0.66
opencv-contrib-python-headless==4.6.0.66
opencv-contrib-python==4.6.0.66
vietocr==0.3.11
ultralytics==8.0.150
Polygon3==3.0.9.1
lanms-neo==1.0.2
numpy==1.24.4
Pillow==9.5.0
pyclipper==1.3.0.post5
premailer==3.10.0
